feat(.github/workflows/debug.yml): add encoding of OpenAPI schema
This commit adds two new jobs in the debug.yml GitHub workflow file: 'Encode OpenAPI schema' and 'Print summary'.
In 'Encode OpenAPI schema', it decodes an API_SPEC secret and saves it to a file, logs the size of secret contents, encodes the file back with base64, and calculates the time taken for this process.
All this information (size and time taken) is set as output parameters.
'Print summary' simply prints this information. This addition enhances visibility and helps debug any issues with the OpenAPI schema. :+1: