feat: Changed Workflow from Context Testing to Test Decode
This commit significantly modifies the GitHub workflow file `debug.yml`.

Earlier, the workflow was named `Context testing` and was designed to dump different contexts like GitHub, job, steps, runner, strategy, matrix into the log. This proved to be less effective than initially hoped, so we've made a change.

The revised workflow, now called `Test Decode`, simplifies the process. Instead of dumping multiple contexts, it primarily focuses on decoding the OpenAPI schema using the secret value named `BIGSECRET`.

This change makes the GitHub workflow much simpler and more focused on a single task. The task will be run whenever there is a push event on the repository 🚀.