feat: Update terminal commands for OpenAPI schema encoding
The terminal commands for both Unix-like systems and Mac were updated in the README.
These changes involve the inclusion of a new package 'json' for prettifying the encoding process 🍃.  The 'json' package is installed globally with npm. After installation, it is used in conjunction with curl to fetch the OpenAPI schema. The '-0' option is applied to the json command to handle the input as a single JSON object. The output is then piped into base64 and xclip/pbcopy (depending on the system) to copy the encoded contents to the clipboard. With this enhancement, the OpenAPI schema can be processed more cleanly 🧹 and efficiently ⚡️.