feat: Add upload artifact step to Github workflow
In our Github workflow debug.yml file, a new step has been added for uploading the decoded file as an artifact. The uploaded file is specifically the `api_spec.yaml` one. Now, this file will be retained as an artifact after each run, aiding in the easier debugging process for the team. This change enhances the investigation of build logs if necessary. 👷‍♂️💼