feat(.github/workflows): add pretty print to api_spec.yaml
A new feature which adds a step in the workflows to implement 'pretty printing' using jq to the api_spec.yaml. This process is aimed at improving the readability and manageability of our yaml files by neatly formatting them. It should be noted that this change also makes sure we upload the prettified yaml file as part of the workflow artifact, not the original one. Hence, the path in 'upload-artifact' action has been modified accordingly. 📚🛠️