feat: Add random exit in debug workflow
In this commit, an addition has been made in the GitHub Actions workflow file (.github/workflows/debug.yml). A new step labelled "Random exit" has been put in place. This step executes a simple run command that generates a random exit status. The script calculates a random integer (either 0 or 1) and uses this as the exit status for the step. It also logs this status for easier debugging. A quirky way to potentially show how workflows respond to different exit codes. 🎲🏁